backend/requirements.txt
fastapi
uvicorn
databricks-sql-connector
python-dotenv
pydantic
numpy
